moved version to core
