+++ core/pkg/version/version.go
